move imagereferencemutator package to only point of use
