+++ pkg/image/apiserver/admission/imagepolicy/internalimagereferencemutators/builds.go
@@ -1,4 +1,4 @@
-package internalversion
+package internalimagereferencemutators
 
 import (
 	"k8s.io/apimachinery/pkg/api/errors"
