+++ pkg/image/apiserver/admission/imagepolicy/internalimagereferencemutators/builds_test.go
@@ -1,4 +1,4 @@
-package internalversion
+package internalimagereferencemutators
 
 import (
 	"reflect"
